added sample data for dummy report construction
these are datasets that I found from Kaggle that are publicly available. These are the initial datasets that we will be working with to mimic the data types that we are looking to have within the app itself. these datasets are categorical data, count data, and "real-time" or date-related data.